+++ packages/go_router/test/go_router_test.dart
@@ -7,13 +7,13 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/src/foundation/diagnostics.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:go_router/go_router.dart';
-import 'package:go_router/src/go_route_match.dart';
-import 'package:go_router/src/go_router_error_page.dart';
-import 'package:go_router/src/typedefs.dart';
+import 'package:go_router/src/delegate.dart';
+import 'package:go_router/src/match.dart';
 import 'package:logging/logging.dart';
+
+import 'test_helpers.dart';
 
 const bool enableLogs = true;
 final Logger log = Logger('GoRouter tests');
@@ -31,8 +31,8 @@ void main() {
                 const HomeScreen()),
       ];
 
-      final GoRouter router = await _router(routes, tester);
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final GoRouter router = await createRouter(routes, tester);
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
       expect(matches.first.fullpath, '/');
       expect(router.screenFor(matches.first).runtimeType, HomeScreen);
@@ -41,13 +41,13 @@ void main() {
     testWidgets('If there is more than one route to match, use the first match',
         (WidgetTester tester) async {
       final List<GoRoute> routes = <GoRoute>[
-        GoRoute(path: '/', builder: _dummy),
-        GoRoute(path: '/', builder: _dummy),
+        GoRoute(path: '/', builder: dummy),
+        GoRoute(path: '/', builder: dummy),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/');
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
       expect(matches.first.fullpath, '/');
       expect(router.screenFor(matches.first).runtimeType, DummyScreen);
@@ -63,11 +63,11 @@ void main() {
       expect(() {
         GoRoute(
           path: '/',
-          builder: _dummy,
+          builder: dummy,
           routes: <GoRoute>[
             GoRoute(
               path: '/foo',
-              builder: _dummy,
+              builder: dummy,
             ),
           ],
         );
@@ -78,11 +78,11 @@ void main() {
       expect(() {
         GoRoute(
           path: '/',
-          builder: _dummy,
+          builder: dummy,
           routes: <GoRoute>[
             GoRoute(
               path: 'foo/',
-              builder: _dummy,
+              builder: dummy,
             ),
           ],
         );
@@ -93,23 +93,23 @@ void main() {
         (WidgetTester tester) async {
       await expectLater(() async {
         final List<GoRoute> routes = <GoRoute>[
-          GoRoute(path: 'foo', builder: _dummy),
+          GoRoute(path: 'foo', builder: dummy),
         ];
-        await _router(routes, tester);
+        await createRouter(routes, tester);
       }, throwsA(isAssertionError));
     });
 
     testWidgets('match no routes', (WidgetTester tester) async {
       final List<GoRoute> routes = <GoRoute>[
-        GoRoute(path: '/', builder: _dummy),
+        GoRoute(path: '/', builder: dummy),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/foo');
       await tester.pumpAndSettle();
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
-      expect(router.screenFor(matches.first).runtimeType, ErrorScreen);
+      expect(router.screenFor(matches.first).runtimeType, TestErrorScreen);
     });
 
     testWidgets('match 2nd top level route', (WidgetTester tester) async {
@@ -124,9 +124,9 @@ void main() {
                 const LoginScreen()),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/login');
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
       expect(matches.first.subloc, '/login');
       expect(router.screenFor(matches.first).runtimeType, LoginScreen);
@@ -152,9 +152,9 @@ void main() {
                 const LoginScreen()),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/login');
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
       expect(matches.first.subloc, '/login');
       expect(router.screenFor(matches.first).runtimeType, LoginScreen);
@@ -175,9 +175,10 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/login/');
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
+      print(matches);
       expect(matches, hasLength(1));
       expect(matches.first.subloc, '/login');
       expect(router.screenFor(matches.first).runtimeType, LoginScreen);
@@ -187,12 +188,12 @@ void main() {
         (WidgetTester tester) async {
       final List<GoRoute> routes = <GoRoute>[
         GoRoute(path: '/profile', redirect: (_) => '/profile/foo'),
-        GoRoute(path: '/profile/:kind', builder: _dummy),
+        GoRoute(path: '/profile/:kind', builder: dummy),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/profile/');
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
       expect(matches.first.subloc, '/profile/foo');
       expect(router.screenFor(matches.first).runtimeType, DummyScreen);
@@ -202,12 +203,12 @@ void main() {
         (WidgetTester tester) async {
       final List<GoRoute> routes = <GoRoute>[
         GoRoute(path: '/profile', redirect: (_) => '/profile/foo'),
-        GoRoute(path: '/profile/:kind', builder: _dummy),
+        GoRoute(path: '/profile/:kind', builder: dummy),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/profile/?bar=baz');
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
       expect(matches.first.subloc, '/profile/foo');
       expect(router.screenFor(matches.first).runtimeType, DummyScreen);
@@ -229,9 +230,9 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/login');
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches.length, 2);
       expect(matches.first.subloc, '/');
       expect(router.screenFor(matches.first).runtimeType, HomeScreen);
@@ -267,9 +268,9 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       {
-        final List<GoRouteMatch> matches = router.routerDelegate.matches;
+        final List<RouteMatch> matches = router.routerDelegate.matches.matches;
         expect(matches, hasLength(1));
         expect(matches.first.fullpath, '/');
         expect(router.screenFor(matches.first).runtimeType, HomeScreen);
@@ -277,7 +278,7 @@ void main() {
 
       router.go('/login');
       {
-        final List<GoRouteMatch> matches = router.routerDelegate.matches;
+        final List<RouteMatch> matches = router.routerDelegate.matches.matches;
         expect(matches.length, 2);
         expect(matches.first.subloc, '/');
         expect(router.screenFor(matches.first).runtimeType, HomeScreen);
@@ -287,7 +288,7 @@ void main() {
 
       router.go('/family/f2');
       {
-        final List<GoRouteMatch> matches = router.routerDelegate.matches;
+        final List<RouteMatch> matches = router.routerDelegate.matches.matches;
         expect(matches.length, 2);
         expect(matches.first.subloc, '/');
         expect(router.screenFor(matches.first).runtimeType, HomeScreen);
@@ -297,7 +298,7 @@ void main() {
 
       router.go('/family/f2/person/p1');
       {
-        final List<GoRouteMatch> matches = router.routerDelegate.matches;
+        final List<RouteMatch> matches = router.routerDelegate.matches.matches;
         expect(matches.length, 3);
         expect(matches.first.subloc, '/');
         expect(router.screenFor(matches.first).runtimeType, HomeScreen);
@@ -342,19 +343,19 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/bar');
-      List<GoRouteMatch> matches = router.routerDelegate.matches;
+      List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(2));
       expect(router.screenFor(matches[1]).runtimeType, Page1Screen);
 
       router.go('/foo/bar');
-      matches = router.routerDelegate.matches;
+      matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(2));
       expect(router.screenFor(matches[1]).runtimeType, FamilyScreen);
 
       router.go('/foo');
-      matches = router.routerDelegate.matches;
+      matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(2));
       expect(router.screenFor(matches[1]).runtimeType, Page2Screen);
     });
@@ -436,7 +437,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/', extra: 1);
       await tester.pump();
       router.push('/login', extra: 2);
@@ -461,10 +462,10 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       const String loc = '/FaMiLy/f2';
       router.go(loc);
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
 
       // NOTE: match the lower case, since subloc is canonicalized to match the
       // path case whereas the location can be any case; so long as the path
@@ -479,15 +480,15 @@ void main() {
         'If there is more than one route to match, use the first match.',
         (WidgetTester tester) async {
       final List<GoRoute> routes = <GoRoute>[
-        GoRoute(path: '/', builder: _dummy),
-        GoRoute(path: '/page1', builder: _dummy),
-        GoRoute(path: '/page1', builder: _dummy),
-        GoRoute(path: '/:ok', builder: _dummy),
+        GoRoute(path: '/', builder: dummy),
+        GoRoute(path: '/page1', builder: dummy),
+        GoRoute(path: '/page1', builder: dummy),
+        GoRoute(path: '/:ok', builder: dummy),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/user');
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
       expect(router.screenFor(matches.first).runtimeType, DummyScreen);
     });
@@ -503,18 +504,18 @@ void main() {
                 const HomeScreen()),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.goNamed('home');
     });
 
     testWidgets('match too many routes', (WidgetTester tester) async {
       final List<GoRoute> routes = <GoRoute>[
-        GoRoute(name: 'home', path: '/', builder: _dummy),
-        GoRoute(name: 'home', path: '/', builder: _dummy),
+        GoRoute(name: 'home', path: '/', builder: dummy),
+        GoRoute(name: 'home', path: '/', builder: dummy),
       ];
 
       await expectLater(() async {
-        await _router(routes, tester);
+        await createRouter(routes, tester);
       }, throwsA(isAssertionError));
     });
 
@@ -527,9 +528,9 @@ void main() {
     testWidgets('match no routes', (WidgetTester tester) async {
       await expectLater(() async {
         final List<GoRoute> routes = <GoRoute>[
-          GoRoute(name: 'home', path: '/', builder: _dummy),
+          GoRoute(name: 'home', path: '/', builder: dummy),
         ];
-        final GoRouter router = await _router(routes, tester);
+        final GoRouter router = await createRouter(routes, tester);
         router.goNamed('work');
       }, throwsA(isAssertionError));
     });
@@ -550,7 +551,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.goNamed('login');
     });
 
@@ -572,7 +573,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.goNamed('login');
     });
 
@@ -605,7 +606,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.goNamed('person',
           params: <String, String>{'fid': 'f2', 'pid': 'p1'});
     });
@@ -636,7 +637,7 @@ void main() {
         ),
       ];
       await expectLater(() async {
-        final GoRouter router = await _router(routes, tester);
+        final GoRouter router = await createRouter(routes, tester);
         router.goNamed('person', params: <String, String>{'fid': 'f2'});
         await tester.pump();
       }, throwsA(isAssertionError));
@@ -672,7 +673,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.goNamed('person',
           params: <String, String>{'fid': 'f2', 'pid': 'p1'});
     });
@@ -687,7 +688,7 @@ void main() {
         ),
       ];
       await expectLater(() async {
-        final GoRouter router = await _router(routes, tester);
+        final GoRouter router = await createRouter(routes, tester);
         router.goNamed('family');
       }, throwsA(isAssertionError));
     });
@@ -702,7 +703,7 @@ void main() {
         ),
       ];
       await expectLater(() async {
-        final GoRouter router = await _router(routes, tester);
+        final GoRouter router = await createRouter(routes, tester);
         router.goNamed('family',
             params: <String, String>{'fid': 'f2', 'pid': 'p1'});
       }, throwsA(isAssertionError));
@@ -733,11 +734,11 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.goNamed('person',
           params: <String, String>{'fid': 'f2', 'pid': 'p1'});
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(router.screenFor(matches.last).runtimeType, PersonScreen);
     });
 
@@ -755,13 +756,13 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       final String loc = router
           .namedLocation('page1', params: <String, String>{'param1': param1});
       log.info('loc= $loc');
       router.go(loc);
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       log.info('param1= ${matches.first.decodedParams['param1']}');
       expect(router.screenFor(matches.first).runtimeType, DummyScreen);
       expect(matches.first.decodedParams['param1'], param1);
@@ -781,12 +782,12 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       final String loc = router.namedLocation('page1',
           queryParams: <String, String>{'param1': param1});
       router.go(loc);
       await tester.pump();
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(router.screenFor(matches.first).runtimeType, DummyScreen);
       expect(matches.first.queryParams['param1'], param1);
     });
@@ -812,7 +813,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester,
+      final GoRouter router = await createRouter(routes, tester,
           redirect: (GoRouterState state) =>
               state.subloc == '/login' ? null : '/login');
 
@@ -844,7 +845,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         routes,
         tester,
         redirect: (GoRouterState state) =>
@@ -875,7 +876,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/dummy');
       await tester.pump();
       expect(router.location, '/login');
@@ -907,7 +908,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/dummy');
       await tester.pump();
       expect(router.location, '/login');
@@ -935,7 +936,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester,
+      final GoRouter router = await createRouter(routes, tester,
           redirect: (GoRouterState state) =>
               state.subloc == '/dummy1' ? '/dummy2' : null);
       router.go('/dummy1');
@@ -944,22 +945,23 @@ void main() {
     });
 
     testWidgets('top-level redirect loop', (WidgetTester tester) async {
-      final GoRouter router = await _router(<GoRoute>[], tester,
+      final GoRouter router = await createRouter(<GoRoute>[], tester,
           redirect: (GoRouterState state) => state.subloc == '/'
               ? '/login'
               : state.subloc == '/login'
                   ? '/'
                   : null);
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
-      expect(router.screenFor(matches.first).runtimeType, ErrorScreen);
-      expect((router.screenFor(matches.first) as ErrorScreen).ex, isNotNull);
-      log.info((router.screenFor(matches.first) as ErrorScreen).ex);
+      expect(router.screenFor(matches.first).runtimeType, TestErrorScreen);
+      expect(
+          (router.screenFor(matches.first) as TestErrorScreen).ex, isNotNull);
+      log.info((router.screenFor(matches.first) as TestErrorScreen).ex);
     });
 
     testWidgets('route-level redirect loop', (WidgetTester tester) async {
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         <GoRoute>[
           GoRoute(
             path: '/',
@@ -973,15 +975,16 @@ void main() {
         tester,
       );
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
-      expect(router.screenFor(matches.first).runtimeType, ErrorScreen);
-      expect((router.screenFor(matches.first) as ErrorScreen).ex, isNotNull);
-      log.info((router.screenFor(matches.first) as ErrorScreen).ex);
+      expect(router.screenFor(matches.first).runtimeType, TestErrorScreen);
+      expect(
+          (router.screenFor(matches.first) as TestErrorScreen).ex, isNotNull);
+      log.info((router.screenFor(matches.first) as TestErrorScreen).ex);
     });
 
     testWidgets('mixed redirect loop', (WidgetTester tester) async {
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         <GoRoute>[
           GoRoute(
             path: '/login',
@@ -993,16 +996,17 @@ void main() {
             state.subloc == '/' ? '/login' : null,
       );
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
-      expect(router.screenFor(matches.first).runtimeType, ErrorScreen);
-      expect((router.screenFor(matches.first) as ErrorScreen).ex, isNotNull);
-      log.info((router.screenFor(matches.first) as ErrorScreen).ex);
+      expect(router.screenFor(matches.first).runtimeType, TestErrorScreen);
+      expect(
+          (router.screenFor(matches.first) as TestErrorScreen).ex, isNotNull);
+      log.info((router.screenFor(matches.first) as TestErrorScreen).ex);
     });
 
     testWidgets('top-level redirect loop w/ query params',
         (WidgetTester tester) async {
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         <GoRoute>[],
         tester,
         redirect: (GoRouterState state) => state.subloc == '/'
@@ -1012,11 +1016,12 @@ void main() {
                 : null,
       );
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
-      expect(router.screenFor(matches.first).runtimeType, ErrorScreen);
-      expect((router.screenFor(matches.first) as ErrorScreen).ex, isNotNull);
-      log.info((router.screenFor(matches.first) as ErrorScreen).ex);
+      expect(router.screenFor(matches.first).runtimeType, TestErrorScreen);
+      expect(
+          (router.screenFor(matches.first) as TestErrorScreen).ex, isNotNull);
+      log.info((router.screenFor(matches.first) as TestErrorScreen).ex);
     });
 
     testWidgets('expect null path/fullpath on top-level redirect',
@@ -1033,7 +1038,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         routes,
         tester,
         initialLocation: '/dummy',
@@ -1055,7 +1060,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         routes,
         tester,
         initialLocation: '/login?from=/',
@@ -1071,7 +1076,7 @@ void main() {
         },
       );
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
       expect(router.screenFor(matches.first).runtimeType, LoginScreen);
     });
@@ -1094,13 +1099,13 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         routes,
         tester,
         initialLocation: loc,
       );
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
       expect(router.screenFor(matches.first).runtimeType, HomeScreen);
     });
@@ -1135,13 +1140,13 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         routes,
         tester,
         initialLocation: '/family/f2/person/p1',
       );
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches.length, 3);
       expect(router.screenFor(matches.first).runtimeType, HomeScreen);
       expect(router.screenFor(matches[1]).runtimeType, FamilyScreen);
@@ -1151,18 +1156,19 @@ void main() {
     });
 
     testWidgets('redirect limit', (WidgetTester tester) async {
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         <GoRoute>[],
         tester,
-        redirect: (GoRouterState state) => '${state.location}+',
+        redirect: (GoRouterState state) => '/${state.location}+',
         redirectLimit: 10,
       );
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
-      expect(router.screenFor(matches.first).runtimeType, ErrorScreen);
-      expect((router.screenFor(matches.first) as ErrorScreen).ex, isNotNull);
-      log.info((router.screenFor(matches.first) as ErrorScreen).ex);
+      expect(router.screenFor(matches.first).runtimeType, TestErrorScreen);
+      expect(
+          (router.screenFor(matches.first) as TestErrorScreen).ex, isNotNull);
+      log.info((router.screenFor(matches.first) as TestErrorScreen).ex);
     });
   });
 
@@ -1183,7 +1189,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         routes,
         tester,
         initialLocation: '/dummy',
@@ -1204,7 +1210,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         routes,
         tester,
         initialLocation: '/dummy',
@@ -1233,7 +1239,7 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         routes,
         tester,
         initialLocation: '/',
@@ -1259,11 +1265,11 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       for (final String fid in <String>['f2', 'F2']) {
         final String loc = '/family/$fid';
         router.go(loc);
-        final List<GoRouteMatch> matches = router.routerDelegate.matches;
+        final List<RouteMatch> matches = router.routerDelegate.matches.matches;
 
         expect(router.location, loc);
         expect(matches, hasLength(1));
@@ -1287,11 +1293,11 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       for (final String fid in <String>['f2', 'F2']) {
         final String loc = '/family?fid=$fid';
         router.go(loc);
-        final List<GoRouteMatch> matches = router.routerDelegate.matches;
+        final List<RouteMatch> matches = router.routerDelegate.matches.matches;
 
         expect(router.location, loc);
         expect(matches, hasLength(1));
@@ -1313,11 +1319,11 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       final String loc = '/page1/${Uri.encodeComponent(param1)}';
       router.go(loc);
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       log.info('param1= ${matches.first.decodedParams['param1']}');
       expect(router.screenFor(matches.first).runtimeType, DummyScreen);
       expect(matches.first.decodedParams['param1'], param1);
@@ -1336,17 +1342,17 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       router.go('/page1?param1=$param1');
 
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(router.screenFor(matches.first).runtimeType, DummyScreen);
       expect(matches.first.queryParams['param1'], param1);
 
       final String loc = '/page1?param1=${Uri.encodeQueryComponent(param1)}';
       router.go(loc);
 
-      final List<GoRouteMatch> matches2 = router.routerDelegate.matches;
+      final List<RouteMatch> matches2 = router.routerDelegate.matches.matches;
       expect(router.screenFor(matches2[0]).runtimeType, DummyScreen);
       expect(matches2[0].queryParams['param1'], param1);
     });
@@ -1357,11 +1363,11 @@ void main() {
           routes: <GoRoute>[
             GoRoute(
               path: '/:id/:blah/:bam/:id/:blah',
-              builder: _dummy,
+              builder: dummy,
             ),
           ],
           errorBuilder: (BuildContext context, GoRouterState state) =>
-              ErrorScreen(state.error!),
+              TestErrorScreen(state.error!),
           initialLocation: '/0/1/2/0/1',
         );
         expect(false, true);
@@ -1371,7 +1377,7 @@ void main() {
     });
 
     testWidgets('duplicate query param', (WidgetTester tester) async {
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         <GoRoute>[
           GoRoute(
             path: '/',
@@ -1387,14 +1393,14 @@ void main() {
         tester,
         initialLocation: '/?id=0&id=1',
       );
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
       expect(matches.first.fullpath, '/');
       expect(router.screenFor(matches.first).runtimeType, HomeScreen);
     });
 
     testWidgets('duplicate path + query param', (WidgetTester tester) async {
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         <GoRoute>[
           GoRoute(
             path: '/:id',
@@ -1410,16 +1416,16 @@ void main() {
 
       router.go('/0?id=1');
       await tester.pump();
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
       expect(matches, hasLength(1));
       expect(matches.first.fullpath, '/:id');
       expect(router.screenFor(matches.first).runtimeType, HomeScreen);
     });
 
     testWidgets('push + query param', (WidgetTester tester) async {
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         <GoRoute>[
-          GoRoute(path: '/', builder: _dummy),
+          GoRoute(path: '/', builder: dummy),
           GoRoute(
             path: '/family',
             builder: (BuildContext context, GoRouterState state) =>
@@ -1444,19 +1450,20 @@ void main() {
       router.push('/person?fid=f2&pid=p1');
       await tester.pump();
       final FamilyScreen page1 =
-          router.screenFor(router.routerDelegate.matches.first) as FamilyScreen;
+          router.screenFor(router.routerDelegate.matches.matches.first)
+              as FamilyScreen;
       expect(page1.fid, 'f2');
 
-      final PersonScreen page2 =
-          router.screenFor(router.routerDelegate.matches[1]) as PersonScreen;
+      final PersonScreen page2 = router
+          .screenFor(router.routerDelegate.matches.matches[1]) as PersonScreen;
       expect(page2.fid, 'f2');
       expect(page2.pid, 'p1');
     });
 
     testWidgets('push + extra param', (WidgetTester tester) async {
-      final GoRouter router = await _router(
+      final GoRouter router = await createRouter(
         <GoRoute>[
-          GoRoute(path: '/', builder: _dummy),
+          GoRoute(path: '/', builder: dummy),
           GoRoute(
             path: '/family',
             builder: (BuildContext context, GoRouterState state) =>
@@ -1481,11 +1488,12 @@ void main() {
       router.push('/person', extra: <String, String>{'fid': 'f2', 'pid': 'p1'});
       await tester.pump();
       final FamilyScreen page1 =
-          router.screenFor(router.routerDelegate.matches.first) as FamilyScreen;
+          router.screenFor(router.routerDelegate.matches.matches.first)
+              as FamilyScreen;
       expect(page1.fid, 'f2');
 
-      final PersonScreen page2 =
-          router.screenFor(router.routerDelegate.matches[1]) as PersonScreen;
+      final PersonScreen page2 = router
+          .screenFor(router.routerDelegate.matches.matches[1]) as PersonScreen;
       expect(page2.fid, 'f2');
       expect(page2.pid, 'p1');
     });
@@ -1515,14 +1523,14 @@ void main() {
         ),
       ];
 
-      final GoRouter router = await _router(routes, tester);
+      final GoRouter router = await createRouter(routes, tester);
       const String fid = 'f1';
       const String pid = 'p2';
       const String loc = '/family/$fid/person/$pid';
 
       router.push(loc);
       await tester.pump();
-      final List<GoRouteMatch> matches = router.routerDelegate.matches;
+      final List<RouteMatch> matches = router.routerDelegate.matches.matches;
 
       expect(router.location, loc);
       expect(matches, hasLength(2));
@@ -1581,8 +1589,8 @@ void main() {
   });
 
   group('GoRouterHelper extensions', () {
-    final GlobalKey<_DummyStatefulWidgetState> key =
-        GlobalKey<_DummyStatefulWidgetState>();
+    final GlobalKey<DummyStatefulWidgetState> key =
+        GlobalKey<DummyStatefulWidgetState>();
     final List<GoRoute> routes = <GoRoute>[
       GoRoute(
         path: '/',
@@ -1787,354 +1795,18 @@ void main() {
         GoRoute(path: '/', builder: (_, __) => const DummyStatefulWidget()),
         GoRoute(
           path: '/error',
-          builder: (_, __) => const GoRouterErrorScreen(null),
+          builder: (_, __) => TestErrorScreen(TestFailure('exception')),
         ),
       ],
       navigatorBuilder: navigatorBuilder,
     );
 
     final GoRouterDelegate delegate = router.routerDelegate;
-    delegate.builderWithNav(
+    delegate.builder.builderWithNav(
       DummyBuildContext(),
-      GoRouterState(router.routeInformationParser,
+      GoRouterState(router.routeConfiguration,
           location: '/foo', subloc: '/bar', name: 'baz'),
       const Navigator(),
     );
   });
-}
-
-Future<GoRouter> createGoRouter(
-  WidgetTester tester, {
-  GoRouterNavigatorBuilder? navigatorBuilder,
-}) async {
-  final GoRouter goRouter = GoRouter(
-    initialLocation: '/',
-    routes: <GoRoute>[
-      GoRoute(path: '/', builder: (_, __) => const DummyStatefulWidget()),
-      GoRoute(
-        path: '/error',
-        builder: (_, __) => const GoRouterErrorScreen(null),
-      ),
-    ],
-    navigatorBuilder: navigatorBuilder,
-  );
-  await tester.pumpWidget(MaterialApp.router(
-      routeInformationProvider: goRouter.routeInformationProvider,
-      routeInformationParser: goRouter.routeInformationParser,
-      routerDelegate: goRouter.routerDelegate));
-  return goRouter;
-}
-
-Widget fakeNavigationBuilder(
-  BuildContext context,
-  GoRouterState state,
-  Widget child,
-) =>
-    child;
-
-class GoRouterNamedLocationSpy extends GoRouter {
-  GoRouterNamedLocationSpy({required List<GoRoute> routes})
-      : super(routes: routes);
-
-  String? name;
-  Map<String, String>? params;
-  Map<String, String>? queryParams;
-
-  @override
-  String namedLocation(
-    String name, {
-    Map<String, String> params = const <String, String>{},
-    Map<String, String> queryParams = const <String, String>{},
-  }) {
-    this.name = name;
-    this.params = params;
-    this.queryParams = queryParams;
-    return '';
-  }
-}
-
-class GoRouterGoSpy extends GoRouter {
-  GoRouterGoSpy({required List<GoRoute> routes}) : super(routes: routes);
-
-  String? myLocation;
-  Object? extra;
-
-  @override
-  void go(String location, {Object? extra}) {
-    myLocation = location;
-    this.extra = extra;
-  }
-}
-
-class GoRouterGoNamedSpy extends GoRouter {
-  GoRouterGoNamedSpy({required List<GoRoute> routes}) : super(routes: routes);
-
-  String? name;
-  Map<String, String>? params;
-  Map<String, String>? queryParams;
-  Object? extra;
-
-  @override
-  void goNamed(
-    String name, {
-    Map<String, String> params = const <String, String>{},
-    Map<String, String> queryParams = const <String, String>{},
-    Object? extra,
-  }) {
-    this.name = name;
-    this.params = params;
-    this.queryParams = queryParams;
-    this.extra = extra;
-  }
-}
-
-class GoRouterPushSpy extends GoRouter {
-  GoRouterPushSpy({required List<GoRoute> routes}) : super(routes: routes);
-
-  String? myLocation;
-  Object? extra;
-
-  @override
-  void push(String location, {Object? extra}) {
-    myLocation = location;
-    this.extra = extra;
-  }
-}
-
-class GoRouterPushNamedSpy extends GoRouter {
-  GoRouterPushNamedSpy({required List<GoRoute> routes}) : super(routes: routes);
-
-  String? name;
-  Map<String, String>? params;
-  Map<String, String>? queryParams;
-  Object? extra;
-
-  @override
-  void pushNamed(
-    String name, {
-    Map<String, String> params = const <String, String>{},
-    Map<String, String> queryParams = const <String, String>{},
-    Object? extra,
-  }) {
-    this.name = name;
-    this.params = params;
-    this.queryParams = queryParams;
-    this.extra = extra;
-  }
-}
-
-class GoRouterPopSpy extends GoRouter {
-  GoRouterPopSpy({required List<GoRoute> routes}) : super(routes: routes);
-
-  bool popped = false;
-
-  @override
-  void pop() {
-    popped = true;
-  }
-}
-
-class GoRouterRefreshStreamSpy extends GoRouterRefreshStream {
-  GoRouterRefreshStreamSpy(
-    Stream<dynamic> stream,
-  )   : notifyCount = 0,
-        super(stream);
-
-  late int notifyCount;
-
-  @override
-  void notifyListeners() {
-    notifyCount++;
-    super.notifyListeners();
-  }
-}
-
-Future<GoRouter> _router(
-  List<GoRoute> routes,
-  WidgetTester tester, {
-  GoRouterRedirect? redirect,
-  String initialLocation = '/',
-  int redirectLimit = 5,
-}) async {
-  final GoRouter goRouter = GoRouter(
-    routes: routes,
-    redirect: redirect,
-    initialLocation: initialLocation,
-    redirectLimit: redirectLimit,
-    errorBuilder: (BuildContext context, GoRouterState state) =>
-        ErrorScreen(state.error!),
-    debugLogDiagnostics: false,
-  );
-  await tester.pumpWidget(
-    MaterialApp.router(
-      routeInformationProvider: goRouter.routeInformationProvider,
-      routeInformationParser: goRouter.routeInformationParser,
-      routerDelegate: goRouter.routerDelegate,
-    ),
-  );
-  return goRouter;
-}
-
-class ErrorScreen extends DummyScreen {
-  const ErrorScreen(this.ex, {Key? key}) : super(key: key);
-  final Exception ex;
-}
-
-class HomeScreen extends DummyScreen {
-  const HomeScreen({Key? key}) : super(key: key);
-}
-
-class Page1Screen extends DummyScreen {
-  const Page1Screen({Key? key}) : super(key: key);
-}
-
-class Page2Screen extends DummyScreen {
-  const Page2Screen({Key? key}) : super(key: key);
-}
-
-class LoginScreen extends DummyScreen {
-  const LoginScreen({Key? key}) : super(key: key);
-}
-
-class FamilyScreen extends DummyScreen {
-  const FamilyScreen(this.fid, {Key? key}) : super(key: key);
-  final String fid;
-}
-
-class FamiliesScreen extends DummyScreen {
-  const FamiliesScreen({required this.selectedFid, Key? key}) : super(key: key);
-  final String selectedFid;
-}
-
-class PersonScreen extends DummyScreen {
-  const PersonScreen(this.fid, this.pid, {Key? key}) : super(key: key);
-  final String fid;
-  final String pid;
-}
-
-class DummyScreen extends StatelessWidget {
-  const DummyScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) => const Placeholder();
-}
-
-Widget _dummy(BuildContext context, GoRouterState state) => const DummyScreen();
-
-extension on GoRouter {
-  Page<dynamic> _pageFor(GoRouteMatch match) {
-    final List<GoRouteMatch> matches = routerDelegate.matches;
-    final int i = matches.indexOf(match);
-    final List<Page<dynamic>> pages =
-        routerDelegate.getPages(DummyBuildContext(), matches).toList();
-    return pages[i];
-  }
-
-  Widget screenFor(GoRouteMatch match) =>
-      (_pageFor(match) as MaterialPage<void>).child;
-}
-
-class DummyBuildContext implements BuildContext {
-  @override
-  bool get debugDoingBuild => throw UnimplementedError();
-
-  @override
-  InheritedWidget dependOnInheritedElement(InheritedElement ancestor,
-      {Object aspect = 1}) {
-    throw UnimplementedError();
-  }
-
-  @override
-  T? dependOnInheritedWidgetOfExactType<T extends InheritedWidget>(
-      {Object? aspect}) {
-    throw UnimplementedError();
-  }
-
-  @override
-  DiagnosticsNode describeElement(String name,
-      {DiagnosticsTreeStyle style = DiagnosticsTreeStyle.errorProperty}) {
-    throw UnimplementedError();
-  }
-
-  @override
-  List<DiagnosticsNode> describeMissingAncestor(
-      {required Type expectedAncestorType}) {
-    throw UnimplementedError();
-  }
-
-  @override
-  DiagnosticsNode describeOwnershipChain(String name) {
-    throw UnimplementedError();
-  }
-
-  @override
-  DiagnosticsNode describeWidget(String name,
-      {DiagnosticsTreeStyle style = DiagnosticsTreeStyle.errorProperty}) {
-    throw UnimplementedError();
-  }
-
-  // @override
-  // TODO(dit): Remove ignore below when flutter 2.11.0-0.0.pre.724 becomes stable
-  // ignore:annotate_overrides
-  void dispatchNotification(Notification notification) {
-    throw UnimplementedError();
-  }
-
-  @override
-  T? findAncestorRenderObjectOfType<T extends RenderObject>() {
-    throw UnimplementedError();
-  }
-
-  @override
-  T? findAncestorStateOfType<T extends State<StatefulWidget>>() {
-    throw UnimplementedError();
-  }
-
-  @override
-  T? findAncestorWidgetOfExactType<T extends Widget>() {
-    throw UnimplementedError();
-  }
-
-  @override
-  RenderObject? findRenderObject() {
-    throw UnimplementedError();
-  }
-
-  @override
-  T? findRootAncestorStateOfType<T extends State<StatefulWidget>>() {
-    throw UnimplementedError();
-  }
-
-  @override
-  InheritedElement?
-      getElementForInheritedWidgetOfExactType<T extends InheritedWidget>() {
-    throw UnimplementedError();
-  }
-
-  @override
-  BuildOwner? get owner => throw UnimplementedError();
-
-  @override
-  Size? get size => throw UnimplementedError();
-
-  @override
-  void visitAncestorElements(bool Function(Element element) visitor) {}
-
-  @override
-  void visitChildElements(ElementVisitor visitor) {}
-
-  @override
-  Widget get widget => throw UnimplementedError();
-}
-
-class DummyStatefulWidget extends StatefulWidget {
-  const DummyStatefulWidget({Key? key}) : super(key: key);
-
-  @override
-  State<DummyStatefulWidget> createState() => _DummyStatefulWidgetState();
-}
-
-class _DummyStatefulWidgetState extends State<DummyStatefulWidget> {
-  @override
-  Widget build(BuildContext context) => Container();
 }